00-playground/61-ai-agents-automation/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.71.0",
    "python-dotenv>=1.1.0",
    "serpapi>=0.1.5",
]
